requirements.txt
appdirs==1.4.4
appnope==0.1.0
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.2.10
attrs==20.1.0
autopep8==1.5.7
backcall==0.2.0
bcrypt==3.1.7
bleach==3.1.5
blinker==1.4
Brlapi==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cliapp==1.20180812.1
Click==7.0
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
DateTime==4.3
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-extensions==3.1.3
django-filter==2.3.0
django-multiselectfield==0.1.12
django-phonenumber-field==5.0.0
django-remote-forms==0.0.1
django-tinymce==3.3.0
django-tinymce4-lite==1.8.0
django-twilio==0.13.0.post1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
duplicity==0.8.12.0
entrypoints==0.3
Faker==4.1.1
fasteners==0.14.1
filelock==3.0.12
frappe-bench==5.4.1
future==0.18.2
geographiclib==1.50
geojson==2.5.0
geopy==2.0.0
gitdb2==2.0.6
GitPython==2.1.15
gmaps==0.9.0
gunicorn==20.1.0
honcho==1.0.1
httplib2==0.18.1
idna==2.10
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsmin==2.2.2
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
Mako==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
mistune==0.8.4
monotonic==1.5
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.4
notebook==6.1.3
numpy==1.19.2
oauthlib==3.1.0
packaging==20.4
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.7.1
pexpect==4.8.0
phonenumbers==8.12.10
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
ptyprocess==0.6.0
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.20
pycups==1.9.73
Pygments==2.6.1
PyGObject==3.36.0
PyJWT==2.1.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.5
python-crontab==2.4.0
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-gmaps==0.3.1
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
semantic-version==2.8.2
Send2Trash==1.5.0
simplejson==3.16.0
six==1.15.0
smmap==4.0.0
smmap2==3.0.1
sqlparse==0.3.1
systemd-python==234
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
toml==0.10.2
tornado==6.0.4
traitlets==4.3.3
ttystatus==0.38
tweepy==3.9.0
twilio==5.7.0
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
usb-creator==0.3.7
virtualenv==20.4.7
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
xkit==0.0.0
zope.interface==5.1.0
